+++ eth_defi/erc_4626/classification.py
@@ -1135,6 +1135,9 @@ HARDCODED_PROTOCOLS = {
     # cSigma Finance - CsigmaV2Pool on Ethereum
     # https://etherscan.io/address/0x438982ea288763370946625fd76c2508ee1fb229
     "0x438982ea288763370946625fd76c2508ee1fb229": {ERC4626Feature.csigma_like},
+    # cSigma Finance - cSuperior Quality Private Credit vault on Ethereum
+    # https://etherscan.io/address/0x50d59b785df23728d9948804f8ca3543237a1495
+    "0x50d59b785df23728d9948804f8ca3543237a1495": {ERC4626Feature.csigma_like},
     # Spark - sUSDC vault on Ethereum
     # https://etherscan.io/address/0xbc65ad17c5c0a2a4d159fa5a503f4992c7b545fe
     "0xbc65ad17c5c0a2a4d159fa5a503f4992c7b545fe": {ERC4626Feature.spark_like},